Moved screenshot download button in BottomDockChart.
